tests/fixtures/factories.py
import factory
from models.user import User

class UserFactory(factory.alchemy.SQLAlchemyModelFactory):
    class Meta:
        model = User
        sqlalchemy_session_persistence = 'commit'

    username = factory.Sequence(lambda n: f'user{n}')
    email = factory.LazyAttribute(lambda o: f'{o.username}@example.com')
    password = 'testpassword'
    active = True
    fs_uniquifier = factory.Faker('uuid4') 