requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-Migrate==4.0.5
Flask-Admin>=1.6.1
Flask-Principal==0.4.0
Flask-RESTX==1.3.0
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
pytest==7.4.0
pytest-flask==1.3.0
sentry-sdk==1.39.1
# htmx is a frontend library, not a Python package, so it will be included via CDN in templates 
Flask-Security-Too==5.3.0
psycopg2-binary>=2.9.0

# Test/automation dependencies
selenium
webdriver-manager
chromedriver-autoinstaller
pytest-xvfb
tenacity
factory-boy